Pass only paths relative to the git repo to the stash tool
Often vi opens files using the full path, but the stash tool provided by the
atlassian-stash gem expects paths passed to the browse command to be relative
to the basepath of the git repository. With these changes the BitBucket
function now changes absolution paths to relative paths.

Note: This implementation depends on fugitive.vim.