Fall back to simplistic fix on save linting via prettier
Removes naive configuration detection and, as a consequence, default to
semistandard. If I were restore this feature, I would rather asking each
tool (e.g. eslint) to determine if a project config is present. Or, if
possible, rely on adding .editorconfig to any future project.

Note to self:

Initially the fix on save behavior was not working, even after
restarting the editor. ALE was detecting an error having to do with what
looks like a buffer file that simply no longer existed (you could see
this by running :ALEInfo), but it was also swallowing it; the failure
was silent.

Should this happen again it may be worth looking into how errors in ALE
might be reported in the quickfix menu, or something like it.