Fix rainbow bracket highlighting
Specifically, enable it for clojure, but also restore the ability to run
the RainbowToggle* commands for any filetype.

Uses rainbow's filetype [configuration] from the luochen1990/rainbow
README to somehow trick the plugin to work. Notable, the
[troubleshooting] advice wasn't needed at all.

Possibly caused by a conflict with another plugin used in this
configuration. Don't know which one.

[configuration]: https://github.com/luochen1990/rainbow#configure
[troubleshooting]: https://github.com/luochen1990/rainbow#troubleshooting